+++ Forge/src/main/java/com/christofmeg/brutalharvest/common/init/EnchantmentRegistry.java
@@ -1,10 +1,12 @@
 package com.christofmeg.brutalharvest.common.init;
 
 import com.christofmeg.brutalharvest.CommonConstants;
+import com.christofmeg.brutalharvest.common.enchantment.BoomerangEnchantment;
 import net.minecraft.world.item.enchantment.Enchantment;
 import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.registries.DeferredRegister;
 import net.minecraftforge.registries.ForgeRegistries;
+import net.minecraftforge.registries.RegistryObject;
 
 import javax.annotation.Nonnull;
 
@@ -16,8 +18,10 @@ public class EnchantmentRegistry {
         ENCHANTMENTS.register(modEventBus);
     }
 
-    static {
+    public static final RegistryObject<Enchantment> BOOMERANG;
 
+    static {
+        BOOMERANG = ENCHANTMENTS.register("boomerang", BoomerangEnchantment::new);
     }
 
 }