+++ Forge/changelog.md
@@ -1,3 +1,4 @@
 1.11.0:
-* Added Scythes
+* Added Throwable Scythes
+* Added Boomerang enchantment
 * Crops can be bone mealed to mature age